+++ plugins/backstage-plugin-opa-entity-checker-processor/src/types.ts
@@ -1,0 +1,17 @@
+export type ValidationResult = {
+  id?: string;
+  check_title?: string;
+  level: 'error' | 'warning' | 'info' | 'success';
+  url?: string;
+  decisionId?: string;
+  message: string;
+};
+
+export type ValidationResponse = {
+  result?: ValidationResult[];
+};
+
+export type OpaConfig = {
+  baseUrl: string;
+  entrypoint: string;
+};
